Fix view template name generation on Windows
On Linux & Mac `Render.getRenderName()` generated a name in the
format `view-subdir/template-file`, while on Windows the generated
name was `view-subdir\template-file`. Since the generated name is
used as an identifier in `Render.Templates` it was impossible
to refer to a template in a cross-platform manner. This change
ensures that the generate template names always contain
forward-slashes, even on Windows.